new file:   100m_2020-2021.npy 	modified:   clean_data.py 	new file:   statistical.py 	renamed:    main copy.py -> trainer.py
